src/requirements.txt
fastapi==0.92.0
python-multipart
aiofiles
uvicorn[standard]
python-dotenv
python-jose[cryptography]


asyncpg==0.27.0
databases[postgresql]==0.6.2
psycopg2-binary==2.9.5
SQLAlchemy==1.4.41

numpy==1.21.3
pandas==1.3.4
onnxruntime==1.7.0
Pillow>=9.2.0
loguru
opencv-python  # need these installed with command: sudo apt-get update && apt-get install ffmpeg libsm6 libxext6  -y
pyyaml